+++ db-tool/src/main/java/com/lizp/constant/Constant.java
@@ -1,0 +1,15 @@
+package com.lizp.constant;
+
+/**
+ * @author lizp4
+ * @date 2022-10-8 15:29
+ * constant
+ */
+public class Constant {
+    public static final String FILE_SEPARATOR = System.getProperty("file.separator");
+    public static final String LINE_SEPARATOR = System.getProperty("line.separator");
+
+    private Constant() {
+        //do nothing
+    }
+}
